Combined some handler classes into a single module
