+++ brew_view/__init__.py
@@ -242,84 +242,53 @@ def _setup_scheduler():
 
 
 def _setup_tornado_app():
-
     # Import these here so we don't have a problem importing thrift_context
-    from brew_view.controllers import (
-        AdminAPI,
-        CommandAPI,
-        CommandListAPI,
-        ConfigHandler,
-        InstanceAPI,
-        QueueAPI,
-        QueueListAPI,
-        RequestAPI,
-        RequestListAPI,
-        SystemAPI,
-        SystemListAPI,
-        VersionHandler,
-        SpecHandler,
-        SwaggerConfigHandler,
-        OldAdminAPI,
-        OldQueueAPI,
-        OldQueueListAPI,
-        LoggingConfigAPI,
-        EventPublisherAPI,
-        EventSocket,
-        TokenAPI,
-        UserAPI,
-        UsersAPI,
-        RoleAPI,
-        RolesAPI,
-        TokenListAPI,
-        JobAPI,
-        JobListAPI,
-        PermissionsAPI,
-    )
+    import brew_view.controllers as bc
 
     prefix = config.web.url_prefix
     static_base = os.path.join(os.path.dirname(__file__), "static", "dist")
 
     # These get documented in our OpenAPI (fka Swagger) documentation
     published_url_specs = [
-        (r"{0}api/v1/commands/?".format(prefix), CommandListAPI),
-        (r"{0}api/v1/requests/?".format(prefix), RequestListAPI),
-        (r"{0}api/v1/systems/?".format(prefix), SystemListAPI),
-        (r"{0}api/v1/queues/?".format(prefix), QueueListAPI),
-        (r"{0}api/v1/users/?".format(prefix), UsersAPI),
-        (r"{0}api/v1/roles/?".format(prefix), RolesAPI),
-        (r"{0}api/v1/permissions/?".format(prefix), PermissionsAPI),
-        (r"{0}api/v1/tokens/?".format(prefix), TokenListAPI),
-        (r"{0}api/v1/admin/?".format(prefix), AdminAPI),
-        (r"{0}api/v1/jobs/?".format(prefix), JobListAPI),
-        (r"{0}api/v1/commands/(\w+)/?".format(prefix), CommandAPI),
-        (r"{0}api/v1/instances/(\w+)/?".format(prefix), InstanceAPI),
-        (r"{0}api/v1/requests/(\w+)/?".format(prefix), RequestAPI),
-        (r"{0}api/v1/systems/(\w+)/?".format(prefix), SystemAPI),
-        (r"{0}api/v1/queues/([\w\.-]+)/?".format(prefix), QueueAPI),
-        (r"{0}api/v1/users/(\w+)/?".format(prefix), UserAPI),
-        (r"{0}api/v1/roles/(\w+)/?".format(prefix), RoleAPI),
-        (r"{0}api/v1/tokens/(\w+)/?".format(prefix), TokenAPI),
-        (r"{0}api/v1/jobs/(\w+)/?".format(prefix), JobAPI),
-        (r"{0}api/v1/config/logging/?".format(prefix), LoggingConfigAPI),
+        (r"{0}api/v1/commands/?".format(prefix), bc.command_api.CommandListAPI),
+        (r"{0}api/v1/requests/?".format(prefix), bc.request_api.RequestListAPI),
+        (r"{0}api/v1/systems/?".format(prefix), bc.system_api.SystemListAPI),
+        (r"{0}api/v1/queues/?".format(prefix), bc.queue_api.QueueListAPI),
+        (r"{0}api/v1/users/?".format(prefix), bc.users_api.UsersAPI),
+        (r"{0}api/v1/roles/?".format(prefix), bc.roles_api.RolesAPI),
+        (r"{0}api/v1/permissions/?".format(prefix), bc.permissions_api.PermissionsAPI),
+        (r"{0}api/v1/tokens/?".format(prefix), bc.token_api.TokenListAPI),
+        (r"{0}api/v1/admin/?".format(prefix), bc.admin_api.AdminAPI),
+        (r"{0}api/v1/jobs/?".format(prefix), bc.job_api.JobListAPI),
+        (r"{0}api/v1/commands/(\w+)/?".format(prefix), bc.command_api.CommandAPI),
+        (r"{0}api/v1/instances/(\w+)/?".format(prefix), bc.instance_api.InstanceAPI),
+        (r"{0}api/v1/requests/(\w+)/?".format(prefix), bc.request_api.RequestAPI),
+        (r"{0}api/v1/systems/(\w+)/?".format(prefix), bc.system_api.SystemAPI),
+        (r"{0}api/v1/queues/([\w\.-]+)/?".format(prefix), bc.queue_api.QueueAPI),
+        (r"{0}api/v1/users/(\w+)/?".format(prefix), bc.users_api.UserAPI),
+        (r"{0}api/v1/roles/(\w+)/?".format(prefix), bc.roles_api.RoleAPI),
+        (r"{0}api/v1/tokens/(\w+)/?".format(prefix), bc.token_api.TokenAPI),
+        (r"{0}api/v1/jobs/(\w+)/?".format(prefix), bc.job_api.JobAPI),
+        (r"{0}api/v1/config/logging/?".format(prefix), bc.logging_api.LoggingConfigAPI),
         # Beta
-        (r"{0}api/vbeta/events/?".format(prefix), EventPublisherAPI),
+        (r"{0}api/vbeta/events/?".format(prefix), bc.event_api.EventPublisherAPI),
         # Deprecated
-        (r"{0}api/v1/admin/system/?".format(prefix), OldAdminAPI),
-        (r"{0}api/v1/admin/queues/?".format(prefix), OldQueueListAPI),
-        (r"{0}api/v1/admin/queues/([\w\.-]+)/?".format(prefix), OldQueueAPI),
+        (r"{0}api/v1/admin/system/?".format(prefix), bc.admin_api.OldAdminAPI),
+        (r"{0}api/v1/admin/queues/?".format(prefix), bc.queue_api.OldQueueListAPI),
+        (r"{0}api/v1/admin/queues/([\w\.-]+)/?".format(prefix), bc.queue_api.OldQueueAPI),
     ]
 
     # And these do not
     unpublished_url_specs = [
         # These are a little special - unpublished but still versioned
         # The swagger spec
-        (r"{0}api/v1/spec/?".format(prefix), SpecHandler),
+        (r"{0}api/v1/spec/?".format(prefix), bc.misc_controllers.SpecHandler),
         # Events websocket
-        (r"{0}api/v1/socket/events/?".format(prefix), EventSocket),
+        (r"{0}api/v1/socket/events/?".format(prefix), bc.event_api.EventSocket),
         # Version / configs
-        (r"{0}version/?".format(prefix), VersionHandler),
-        (r"{0}config/?".format(prefix), ConfigHandler),
-        (r"{0}config/swagger/?".format(prefix), SwaggerConfigHandler),
+        (r"{0}version/?".format(prefix), bc.misc_controllers.VersionHandler),
+        (r"{0}config/?".format(prefix), bc.misc_controllers.ConfigHandler),
+        (r"{0}config/swagger/?".format(prefix), bc.misc_controllers.SwaggerConfigHandler),
         # Not sure if these are really necessary
         (r"{0}".format(prefix[:-1]), RedirectHandler, {"url": prefix}),
         (
